Address PR review feedback: vectorize kinematic target writes, improve logging
- Replace CPU-bound Python for-loop in write_nodal_kinematic_target_to_sim_index
  with vectorized torch advanced indexing (eliminates per-env GPU-CPU sync).
- Log sync_particles_to_usd failures at warning level on first occurrence,
  then throttle to debug for subsequent failures.
- Cache usdrt module import to avoid re-importing every render frame.
- Add clarifying comment on particle_q_prev aliasing in two-way coupling
  friction (known limitation, acceptable for one-way primary use case).
- Document _MAX_REACTION_CONTACTS fixed upper bound in coupled_solver.
